+++ CHANGES.md
@@ -5,6 +5,7 @@ Changes in 12.1.2
 -----------------
 - SQL fix last and first name were mixed when importing user/student from Moodle in ImportUsers.fnc.php
 - Fix PHP warning when updating course period teacher in Moodle/functions.php & Moodle/Scheduling/Courses.php
+- Fix enrol teacher into course instead of just assigning teacher role in Moodle/functions.php, Moodle/Scheduling/Courses.php & Moodle/School_Setup/Rollover.php
 - MySQL fix You can't specify target table 'ac' for update in FROM clause in Courses.fnc.php
 
 Changes in 12.1.1